fix: Add the missing feature flag for `ewm_mean_by`
